backend/cert/requirements.txt
fastapi==0.75.0
uvicorn==0.17.0
pymongo==3.12.1
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-api-python-client==2.29.0
pdfminer.six==20201018
python-dotenv==0.19.2
